// deps/mpl-core/js/src/generated/types/registryRecord.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getStructDecoder,
  getStructEncoder,
  getU64Decoder,
  getU64Encoder,
  type Codec,
  type Decoder,
  type Encoder,
} from '@solana/kit';
import {
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getPluginTypeDecoder,
  getPluginTypeEncoder,
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type PluginType,
  type PluginTypeArgs,
} from '.';

export type RegistryRecord = {
  pluginType: PluginType;
  authority: BasePluginAuthority;
  offset: bigint;
};

export type RegistryRecordArgs = {
  pluginType: PluginTypeArgs;
  authority: BasePluginAuthorityArgs;
  offset: number | bigint;
};

export function getRegistryRecordEncoder(): Encoder<RegistryRecordArgs> {
  return getStructEncoder([
    ['pluginType', getPluginTypeEncoder()],
    ['authority', getBasePluginAuthorityEncoder()],
    ['offset', getU64Encoder()],
  ]);
}

export function getRegistryRecordDecoder(): Decoder<RegistryRecord> {
  return getStructDecoder([
    ['pluginType', getPluginTypeDecoder()],
    ['authority', getBasePluginAuthorityDecoder()],
    ['offset', getU64Decoder()],
  ]);
}

export function getRegistryRecordCodec(): Codec<
  RegistryRecordArgs,
  RegistryRecord
> {
  return combineCodec(getRegistryRecordEncoder(), getRegistryRecordDecoder());
}